Utils.cpp : Do not check for sdcardfs in IsSdcardfsUsed
In order to return true, IsSdcardfsUsed have 2 conditions : external_storage.sdcardfs.enabled should not be 0 and sdcardfs have to available on /proc/filesystem. The thing is, sdcard command support esdfs bind mounting when ro.sys.sdcardfs set to false. So in order to have it bind mount you will have to compile both sdcardfs and esdfs at the same time.

Remove the sdcardfs check and only use the property to check if sdcardfs is being used. This will help us to use esdfs without having to compile sdcardfs

Change-Id: I474150213b3f35361bc18623d84bfdd83ef704d6